WIP: refactor to use builder
